Split Option subcomponents
