+++ packages/fuselage/src/components/Options/Option/OptionAvatar.tsx
@@ -1,0 +1,7 @@
+import React, { FC } from 'react';
+
+const OptionAvatar: FC = (props) => (
+  <div className='rcx-option__avatar' {...props} />
+);
+
+export default OptionAvatar;
