+++ packages/fuselage/src/components/Options/Option/OptionColumn.tsx
@@ -1,0 +1,7 @@
+import React, { FC } from 'react';
+
+const OptionColumn: FC = (props) => (
+  <div className='rcx-option__column' {...props} />
+);
+
+export default OptionColumn;
