+++ packages/fuselage/src/components/Options/Option/OptionContent.tsx
@@ -1,0 +1,7 @@
+import React, { FC } from 'react';
+
+const OptionContent: FC = (props) => (
+  <div className='rcx-option__content' {...props} />
+);
+
+export default OptionContent;
