+++ packages/fuselage/src/components/Options/Option/OptionDescription.tsx
@@ -1,0 +1,7 @@
+import React, { FC } from 'react';
+
+const OptionDescription: FC = (props) => (
+  <div className='rcx-option__description' {...props} />
+);
+
+export default OptionDescription;
