+++ packages/fuselage/src/components/Options/Option/OptionMenu.tsx
@@ -1,0 +1,7 @@
+import React, { FC } from 'react';
+
+const OptionMenu: FC = (props) => (
+  <div className='rcx-box--animated rcx-option__menu-wraper' {...props} />
+);
+
+export default OptionMenu;
