+++ packages/fuselage/src/components/Options/index.js
@@ -3,7 +3,7 @@ import React, { useLayoutEffect, useMemo } from 'react';
 import { Box, Scrollable } from '../Box';
 import { CheckBox } from '../CheckBox';
 import Tile from '../Tile';
-import { Option } from './Option';
+import Option from './Option';
 import { useCursor } from './useCursor';
 
 export { useCursor };
