+++ packages/fuselage/src/components/OptionsPaginated/OptionsPaginated.stories.mdx
@@ -2,7 +2,7 @@ import { Meta, Canvas, Story } from '@storybook/addon-docs';
 
 import { OptionsPaginated, CheckOption } from '.';
 import { Box } from '..';
-import { Option } from '../Options/Option';
+import Option from '../Options/Option';
 
 export const options = [
   { value: 1, label: 'a teste 1', selected: true },
