+++ packages/fuselage/src/components/OptionsPaginated/index.js
@@ -7,7 +7,7 @@ import { Virtuoso } from 'react-virtuoso';
 
 import { AnimatedVisibility, Box } from '../Box';
 import { CheckBox } from '../CheckBox';
-import { Option } from '../Options/Option';
+import Option from '../Options/Option';
 import Tile from '../Tile';
 
 const prevent = (e) => {
@@ -74,7 +74,7 @@ export const OptionsPaginated = React.forwardRef(
               // TODO Add a scroller element
               // components={{ Scroller: ScrollableContentWrapper }}
               itemContent={(index, data) => (
-                <OptionsComponentWithData index={index} data={data} ref={ref} />
+                <OptionsComponentWithData index={index} data={data} />
               )}
             />
           )}
